Home 🏠

Functions pkg_template unvendor vendor armadillo_version

Vignettes Basic 'armadillo4r' usage Armadillo configuration Decompositions, factorisations, inverses and equation solvers (dense matrices) Decompositions, factorisations, inverses and equation solvers (sparse matrices) Functions of vectors, matrices, and cubes Generated vectors, matrices, and cubes Fitting regressions with Armadillo Matrix, vector, cube and field classes Member functions and variables Miscellaneous Saving and loading Armadillo objects on C++ side Signal and image processing Sparse matrices Statistics and clustering Syntax comparison for MATLAB/Octave users Using SEXP and cpp conversions

News Changelog

pkg_template {armadillo4r}	R Documentation

Start a new project with the armadillo4r package template

Description

Start a new project with the armadillo4r package template

Usage

pkg_template(path = NULL, pkgname = NULL)

Arguments

path 	 Path to the new project

pkgname 	 Name of the new package

Value

The file path to the copied template (invisibly).

Examples

# create a new directory
dir <- tempdir()
dir.create(dir)

# copy the package template into the directory
pkg_template(dir, "mynewpkg")