Home 🏠

Functions pkg_template unvendor vendor armadillo_version

Vignettes Basic 'armadillo4r' usage Armadillo configuration Converting between R and Armadillo types Decompositions, factorisations, inverses and equation solvers (dense matrices) Decompositions, factorisations, inverses and equation solvers (sparse matrices) Functions of vectors, matrices, and cubes Generated vectors, matrices, and cubes Fitting regressions with Armadillo Matrix, vector, cube and field classes Member functions and variables Miscellaneous Saving and loading Armadillo objects on C++ side Signal and image processing Sparse matrices Statistics and clustering Syntax comparison for MATLAB/Octave users Using SEXP with Armadillo data types

News Changelog

unvendor {armadillo4r}	R Documentation

Unvendor the armadillo4r headers

Description

This function removes the vendored armadillo4r headers from your package by automatically finding the vendored headers.

Usage

unvendor(path = NULL)

Arguments

path 	 The directory with the vendored headers. It is recommended to use '"./src/vendor"'. The default is 'NULL'.

Value

The path to the unvendored code (invisibly).

Examples

# create a new directory
dir <- paste0(tempdir(), "/", gsub("\\s+|[[:punct:]]", "", Sys.time()))
dir.create(dir, recursive = TRUE)

# vendor the armadillo4r headers into the directory
vendor(dir)

# unvendor the armadillo4r headers from the directory
unvendor(dir)

# cleanup
unlink(dir, recursive = TRUE)