Home 🏠

Functions pkg_template unvendor vendor armadillo_version

Vignettes 01 - Basic 'armadillo4r' usage 02 - Armadillo configuration 03 - Converting between R and Armadillo types 04 - Syntax comparison for MATLAB/Octave users 05 - Sparse matrices 06 - Matrix, vector, cube and field classes 07 - Member functions and variables 08 - Functions of vectors, matrices, and cubes 09 - Generated vectors, matrices, and cubes 10 - Decompositions, factorisations, inverses and equation solvers (dense matrices) 11 - Decompositions, factorisations, inverses and equation solvers (sparse matrices) 12 - Statistics and clustering 13 - Signal and image processing 14 - Saving and loading Armadillo objects on C++ side 15 - Using SEXP with Armadillo data types 16 - Fitting regressions with Armadillo 17 - Miscellaneous

News Changelog

vendor {armadillo4r}	R Documentation

Vendor the cpp4r and armadillo4r headers

Description

Vendors armadillo4r into your package by copying the armadillo4r headers into the inst/include folder of your package and adding 'armadillo4r version: XYZ' to the top of the files, where XYZ is the version of armadillo4r currently installed on your machine.

Note: vendoring places the responsibility of updating the code on you. Bugfixes and new features in armadillo4r will not be available for your code until you run cpp_vendor() again.

Usage

vendor(path = NULL)

Arguments

path 	 The directory with the vendored headers. It is recommended to use "./src/vendor". The default is NULL.

Value

The path to the vendored code (invisibly).

Examples

# create a new directory
dir <- paste0(tempdir(), "/", gsub("\\s+|[[:punct:]]", "", Sys.time()))
dir.create(dir, recursive = TRUE)

# vendor the armadillo4r headers into the directory
vendor(dir)